+++ env.d.ts
@@ -1,1 +1,15 @@
 /// <reference types="vite/client" />
+declare module '*.vue' {
+  import { defineComponent } from 'vue'
+  const Component: ReturnType<typeof defineComponent>
+  export default Component
+}
+
+declare global {
+  namespace JSX {
+    interface IntrinsicAttributes {
+      class?: unknown
+      style?: unknown
+    }
+  }
+}