+++ vitest.config.ts
@@ -7,8 +7,17 @@ export default mergeConfig(
   defineConfig({
     test: {
       environment: 'jsdom',
-      exclude: [...configDefaults.exclude, 'e2e/**', 'component/**'],
+      exclude: [...configDefaults.exclude],
       root: fileURLToPath(new URL('./', import.meta.url)),
+      coverage: {
+        enabled: true,
+        exclude: ['**/play/**', '**/dist/**', '**/theme/**'],
+        excludeAfterRemap: true,
+        include: ['**/*.{js,ts,vue}'],
+        provider: 'v8', // 使用 v8 覆盖率
+        reportsDirectory: './coverage',
+        reporter: ['text', 'html']
+      },
     },
   }),
 )